% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMVC.R
\name{IMVCS}
\alias{IMVCS}
\title{Integrated Mean Variance Correlation Based Screening}
\usage{
IMVCS(y, x, K, d, NN = 3, type)
}
\arguments{
\item{y}{is the response vector}

\item{x}{is the covariate matrix}

\item{K}{is the number of quantile levels}

\item{d}{is the size of selected variables}

\item{NN}{is the number of B spline basis, default is 3}

\item{type}{is an indicator for measuring linear or nonlinear correlation, "linear" represents linear correlation and "nonlinear" represents linear or nonlinear correlation using B splines}
}
\description{
This function is used to select important features using integrated mean variance correlation
}
